LPS-54457 Move bootstrap classes used by hibernate from com.liferay.portal.cache.bootstrap to com.liferay.portal.dao.orm.hibernate.region
