+++ portal-impl/src/com/liferay/portal/cache/ehcache/EhcacheConfigurationUtil.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.cache.ehcache;
 
-import com.liferay.portal.cache.bootstrap.EhcacheStreamBootstrapCacheLoaderFactory;
 import com.liferay.portal.cache.cluster.EhcachePortalCacheClusterReplicatorFactory;
+import com.liferay.portal.dao.orm.hibernate.region.EhcacheStreamBootstrapCacheLoaderFactory;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.util.PropsValues;
